requirements.txt
flask==2.3.2
werkzeug==2.3.6
azure-cognitiveservices-vision-computervision==0.9.0
msrest==0.6.21
